functions_man.py_added
